refactor: move VolunteerDashboard out of RoleDashboards.tsx
Extracts VolunteerDashboard into its own file,
frontend/src/components/dashboards/VolunteerDashboard.tsx, following
the same pattern already established by RegionalAnalyticsView's split.
RoleDashboards.tsx now imports and re-exports it so all existing call
sites (App.tsx etc.) are unaffected.

Mechanical extraction only, no behavior change: verified byte-for-byte
against the original function body, tsc --noEmit clean, and a full
vite build succeeds.

Closes #142
